Add jobby output to console log
For some time I was in need to use jobby with console output. Since
we are working in cloud foundry environment and we capture application
log with a drainlog this was quite critical.

This commit introduces the config flag `output_console` which takes
a boolean to either activate the output to console. What it actually
does is remove the bash output to logs and puts an echo infront of
the `exec()`.

While at it: Add composer script for tests. Add 2 testcases for
closures and files.